ADD: `erode_box` kernel and test
ADD: `dilate_box` kernel and test
ADD: `binary_edge_detection` kernel and test
ADD: `minimum_images` kernel and test
ADD: `maximum_images` kernel and test
ADD: `minimum_sphere` kernel and test
ADD: `maximum_sphere` kernel and test
ADD: `divide_image_and_scalar` kernel and test (Close #158 )
FIX: `InValue` method called twice in replace_intensity
FIX: parameter order for `histogram` and `mask_voronoi_labeling`
FIX: missing test divide_images
FIX: wrong display of `MemoryType` and `DataType` when print

+++ clic/include/tier1/cleMaximumBoxKernel.hpp
@@ -30,9 +30,9 @@ inline auto
 MaximumBoxKernel_Call(const std::shared_ptr<cle::Processor> & device,
                       const Image &                           src,
                       const Image &                           dst,
-                      const float &                           radius_x,
-                      const float &                           radius_y,
-                      const float &                           radius_z) -> void
+                      const int &                             radius_x,
+                      const int &                             radius_y,
+                      const int &                             radius_z) -> void
 {
   MaximumBoxKernel kernel(device);
   kernel.SetInput(src);
